@@ -21,138 +19,95 @@
     "url": "https://github.com/stoplightio/spectral"
   },
   "license": "Apache-2.0",
-  "main": "dist/index.js",
-  "bin": {
-    "spectral": "./dist/cli/index.js"
-  },
-  "files": [
-    "dist/*",
-    "rulesets/*"
-  ],
   "engines": {
     "node": ">=12"
   },
   "scripts": {
-    "build.binary": "pkg . --output ./binaries/spectral",
-    "build.clean": "rimraf ./coverage && rimraf ./dist && rimraf ./rulesets && rimraf ./__karma__/__fixtures__",
-    "build.functions": "rollup -c",
-    "build": "tsc -p ./tsconfig.build.json",
-    "cli": "node -r ts-node/register -r tsconfig-paths/register src/cli/index.ts",
-    "cli:debug": "node -r ts-node/register -r tsconfig-paths/register --inspect-brk src/cli/index.ts",
-    "generate-assets": "ts-node --compiler-options {\\\"module\\\":\\\"CommonJS\\\"} ./scripts/generate-assets.ts",
-    "inline-version": "node ./scripts/inline-version.mjs",
-    "lint.changelog": "kacl lint",
+    "clean": "rimraf packages/*/{dist,.cache}",
+    "prebuild": "yarn workspace @stoplight/spectral-ruleset-migrator prebuild",
+    "build": "tsc --build ./tsconfig.build.json",
+    "postbuild": "yarn workspace @stoplight/spectral-cli postbuild",
+    "lint": "yarn lint.prettier && yarn lint.eslint",
     "lint.fix": "yarn lint.prettier --write && yarn lint.eslint --fix",
-    "lint.eslint": "eslint --cache --cache-location .cache/ --ext=.js,.ts src test-harness scripts",
-    "lint.prettier": "prettier --ignore-path .eslintignore --check {src,test-harness}/**/*.{json,yaml} docs/**/*.md README.md",
-    "lint": "yarn lint.prettier && yarn lint.eslint && yarn lint.changelog",
-    "copy.html-templates": "copyfiles -u 1 \"./src/cli/formatters/html/*.html\" \"./dist/\"",
-    "postbuild.functions": "copyfiles -u 1 \"dist/rulesets/{oas,asyncapi}/functions/*.js\" ./",
-    "postbuild": "yarn build.functions && yarn generate-assets",
-    "prebuild": "yarn build.clean && copyfiles -u 1 \"src/rulesets/{oas,asyncapi}/**/*.json\" dist && copyfiles -u 1 \"src/rulesets/{oas,asyncapi}/**/*.json\" ./ && yarn copy.html-templates",
-    "prebuild.binary": "yarn build",
-    "pretest.karma": "node ./scripts/generate-karma-fixtures.mjs && yarn pretest",
-    "pretest": "yarn generate-assets",
+    "lint.eslint": "eslint --cache --cache-location .cache/ --ext=.js,.mjs,.ts packages test-harness",
+    "lint.prettier": "prettier --ignore-path .eslintignore --ignore-unknown --check packages/core/src/meta/*.json docs/**/*.md README.md",
+    "pretest": "yarn workspace @stoplight/spectral-ruleset-migrator pretest",
+    "test": "yarn test.karma && yarn test.jest --coverage --maxWorkers=2",
     "test.harness": "jest -c ./test-harness/jest.config.js",
-    "test.karma": "karma start",
-    "test.prod": "yarn build && yarn lint && yarn test.karma && yarn test.jest --coverage --maxWorkers=2",
-    "test.update": "yarn test --updateSnapshot",
-    "test.watch": "yarn test --watch",
     "test.jest": "jest --silent",
-    "prepare": "husky install"
+    "test.karma": "karma start",
+    "prepare": "husky install",
+    "postinstall": "patch-package",
+    "release": "HUSKY=0 yarn workspaces run semantic-release -e semantic-release-monorepo"
   },
-  "dependencies": {
-    "@stoplight/better-ajv-errors": "0.1.1",
-    "@stoplight/json": "3.13.5",
-    "@stoplight/json-ref-readers": "1.2.2",
-    "@stoplight/json-ref-resolver": "3.1.2",
-    "@stoplight/lifecycle": "2.3.2",
-    "@stoplight/path": "1.3.2",
-    "@stoplight/types": "12.3.0",
-    "@stoplight/yaml": "4.2.2",
-    "abort-controller": "3.0.0",
-    "ajv": "8.5.0",
-    "ajv-errors": "3.0.0",
-    "ajv-formats": "2.1.0",
-    "blueimp-md5": "2.18.0",
-    "chalk": "4.1.0",
-    "eol": "0.9.1",
-    "expression-eval": "4.0.0",
-    "fast-glob": "3.2.5",
-    "json-schema-migrate": "2.0.0",
-    "json-schema-traverse": "1.0.0",
-    "jsonpath-plus": "5.0.7",
-    "lodash": "4.17.21",
-    "nanoid": "2.1.11",
-    "nimma": "0.0.0",
-    "node-fetch": "2.6.1",
-    "proxy-agent": "4.0.1",
-    "strip-ansi": "6.0",
-    "text-table": "0.2",
-    "tslib": "2.2.0",
-    "yargs": "17.0.1"
+  "workspaces": {
+    "packages": [
+      "packages/*"
+    ]
+  },
+  "release": {
+    "branches": [
+      {
+        "name": "develop",
+        "prerelease": false
+      }
+    ],
+    "plugins": [
+      "@semantic-release/commit-analyzer",
+      "@semantic-release/release-notes-generator",
+      "@semantic-release/npm",
+      "@semantic-release/github",
+      "@semantic-release/git"
+    ]
   },
   "devDependencies": {
-    "@apidevtools/json-schema-ref-parser": "^9.0.7",
-    "@brightcove/kacl": "^0.1.11",
-    "@commitlint/cli": "^12.0.1",
-    "@commitlint/config-conventional": "^12.0.1",
-    "@rollup/plugin-commonjs": "^17.1.0",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^11.2.1",
-    "@types/jest": "^26.0.20",
-    "@types/jest-when": "^2.7.2",
-    "@types/lodash": "^4.14.167",
-    "@types/nanoid": "^2.1.0",
-    "@types/node": "~13.1",
-    "@types/node-fetch": "^2.5.8",
+    "@commitlint/cli": "^13.2.0",
+    "@commitlint/config-conventional": "^12.1.4",
+    "@semantic-release/git": "^9.0.1",
+    "@types/jest": "^27.0.2",
+    "@types/jest-when": "^2.7.3",
+    "@types/karma": "^6.3.1",
+    "@types/lodash": "^4.14.175",
+    "@types/node": "^15.12.4",
+    "@types/node-fetch": "^2.5.12",
     "@types/node-powershell": "^3.1.1",
-    "@types/text-table": "^0.2.1",
-    "@types/tmp": "^0.2.0",
-    "@types/xml2js": "^0.4.8",
-    "@types/yargs": "^16.0.1",
-    "@typescript-eslint/eslint-plugin": "^4.24.0",
-    "@typescript-eslint/parser": "^4.24.0",
-    "copyfiles": "^2.4.1",
-    "dependency-graph": "^0.10.0",
-    "eslint": "^7.27.0",
+    "@types/text-table": "^0.2.2",
+    "@types/tmp": "^0.2.1",
+    "@typescript-eslint/eslint-plugin": "^4.33.0",
+    "@typescript-eslint/parser": "^4.33.0",
+    "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-prettier": "^4.0.0",
+    "fast-glob": "^3.2.5",
     "fetch-mock": "^9.11.0",
-    "glob-fs": "^0.1.7",
-    "husky": "^5.1.3",
-    "jest": "^26.6.3",
-    "jest-mock": "^26.6.2",
-    "jest-when": "^3.2.1",
+    "husky": "^7.0.2",
+    "jest": "^27.2.4",
+    "jest-mock": "^27.2.4",
+    "jest-when": "^3.4.0",
+    "json-schema": "^0.3.0",
     "karma": "^6.1.1",
     "karma-chrome-launcher": "^3.1.0",
     "karma-env-preprocessor": "^0.1.1",
     "karma-jasmine": "^3.3.1",
-    "karma-typescript": "^5.4.0",
-    "karma-typescript-es6-transform": "^5.2.0",
-    "lint-staged": "^10.5.4",
-    "nock": "~12.0.2",
-    "node-html-parser": "^3.3.4",
+    "karma-typescript": "^5.5.2",
+    "karma-typescript-es6-transform": "^5.5.2",
+    "lint-staged": "^11.2.0",
+    "memfs": "^3.3.0",
     "node-powershell": "^4.0.0",
-    "pkg": "^5.1.0",
-    "prettier": "^2.2.1",
-    "recast": "^0.20.4",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.40.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.30.0",
-    "ts-jest": "^26.5.6",
-    "ts-node": "^10.0.0",
-    "tsconfig-paths": "^3.9.0",
-    "typescript": "^4.2.4",
-    "xml2js": "^0.4.23"
+    "patch-package": "^6.4.7",
+    "prettier": "^2.4.1",
+    "semantic-release": "^17.4.4",
+    "semantic-release-monorepo": "^7.0.5",
+    "ts-jest": "^27.0.5",
+    "ts-node": "^10.2.1",
+    "typescript": "^4.3.5"
   },
   "lint-staged": {
-    "*.{ts,tsx}": [
+    "*.{ts,js}": [
       "eslint --fix --cache --cache-location .cache"
     ],
     "CHANGELOG.md": [
-      "kacl lint",
       "prettier --write"
     ],
     "docs/**/*.md": [
